HOTFIX: 멘토 응답 dto 오류 수정

+++ RevUp-domain/src/main/java/com/revup/feedback/service/response/MentorResponse.java
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Data
@@ -24,7 +25,7 @@ public class MentorResponse {
 
     Set<SkillStack> skillStacks;
 
-    private double answerAcceptanceRate;
+    private BigDecimal answerAcceptanceRate;
 
     private Integer liveFeedbackCount;
